The frequency settings screens now pay attention to the data in memory, draw it, load it, and save it. Next do it ofr waveform and effect screens.

+++ Theremin/Model/VWWThereminInputAxis.h
@@ -11,8 +11,8 @@
 #import "VWWThereminSynthesizerSettings.h"
 
 @interface VWWThereminInputAxis : NSObject
-@property (nonatomic) NSUInteger frequencyMax;
-@property (nonatomic) NSUInteger frequencyMin;
+@property (nonatomic) float frequencyMax;
+@property (nonatomic) float frequencyMin;
 @property (nonatomic) WaveType waveType;
 @property (nonatomic) float sensitivity;
 @property (nonatomic) EffectType effectType;
